% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_and_multiply.R
\name{scale_and_multiply}
\alias{scale_and_multiply}
\title{Scale and Multiply}
\usage{
scale_and_multiply(x, multiplier)
}
\arguments{
\item{x}{A numeric vector to be scaled and multiplied, we name it x, representing input values like f(x) = y}

\item{multiplier}{A numeric value by which to multiply the scaled vector, we name it directly as multiplier}
}
\value{
A numeric vector with scaled and multiplied values.
}
\description{
This function scales a numeric vector to the range \verb{[0, 1]} and then multiplies each element by a specified multiplier.
It can be called by: scale_and_multiply(x, multiplier)
}
\examples{
scale_and_multiply(c(1, 2, 3, 4), 10)
scale_and_multiply(c(10, 20, 30), 0.5)

}
